# pet-services-api/internal/transport/http/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an API for managing restaurants.",
        "title": "Pedido Simples API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Pedido Simples",
            "url": "http://www.pedidosimples.com.br",
            "email": "contato@pedidosimples.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/users/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Trocar senha do usuário",
                "parameters": [
                    {
                        "description": "Dados para troca de senha",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.changePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da operação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/users/email/verification/confirm": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Confirmar verificação de email",
                "parameters": [
                    {
                        "description": "Token de verificação",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.confirmEmailVerificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da operação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/users/email/verification/request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Solicitar verificação de email",
                "responses": {
                    "200": {
                        "description": "Status da operação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pet-services-api_internal_domain_user.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Atualizar perfil do usuário",
                "parameters": [
                    {
                        "description": "Dados do perfil",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da operação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Deletar conta do usuário",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Hard delete",
                        "name": "hard",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da operação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/users/password-reset/confirm": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Confirmar reset de senha",
                "parameters": [
                    {
                        "description": "Dados para reset de senha",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.confirmPasswordResetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da operação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/users/password-reset/request": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Solicitar reset de senha",
                "parameters": [
                    {
                        "description": "Email para reset de senha",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.requestPasswordResetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da operação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "exceptions.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "exceptions.ProblemDetailsOutputDTO": {
            "type": "object",
            "properties": {
                "problem_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/exceptions.ProblemDetails"
                    }
                }
            }
        },
...
        "http.changePasswordRequest": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string",
                    "minLength": 1
                },
                "new_password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                }
            }
        },
        "http.confirmEmailVerificationRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "minLength": 10
                }
            }
        },
        "http.confirmPasswordResetRequest": {
            "type": "object",
            "required": [
                "new_password",
                "token"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                },
                "token": {
                    "type": "string",
                    "minLength": 10
                }
            }
        },
...
        "http.requestPasswordResetRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "http.updateProfileRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/user.Address"
                },
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 10
                }
            }
        },
...
        "pet-services-api_internal_domain_user.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "$ref": "#/definitions/user.Email"
                },
                "emailVerified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/user.Location"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "description": "hash bcrypt",
                    "type": "string"
                },
                "phone": {
                    "$ref": "#/definitions/user.Phone"
                },
                "type": {
                    "$ref": "#/definitions/user.UserType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "user.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
        "user.Email": {
            "type": "object"
        },
        "user.Location": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/user.Address"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "user.Phone": {
            "type": "object"
        },
        "user.UserType": {
            "type": "string",
            "enum": [
                "owner",
                "provider"
            ],
            "x-enum-varnames": [
                "UserTypeOwner",
                "UserTypeProvider"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}